bench/requirements.txt
redis>=5.0
paho-mqtt>=2.1
httpx>=0.27
typer>=0.12
rich>=13.7
numpy>=2.0
